requirements.txt
torch
sentencepiece
transformers==4.33
datasets
sacremoses
sacrebleu
unbabel-comet